Remove legacy integration code
