+++ console/agent_settings/mutations.py
@@ -1,0 +1,1 @@
+from .service import handle_agent_settings_mutation
