--- frontend/src/components/agentSettings/settingsSurfaceClasses.ts
@@ -1,6 +1,0 @@
-export {
-  embeddedSettingsSurfaceClassName,
-  getSettingsSurfaceClassName,
-  sharedSettingsGlassFrameClassName,
-  standaloneSettingsSurfaceClassName,
-} from '../common/SettingsSurface'
