+++ frontend/src/screens/AgentDetailScreen.tsx
@@ -1130,6 +1130,24 @@ const toggleOrganizationServer = useCallback((serverId: string) => {
     [applyPeerLinkPayload, submitFormData],
   )
 
+  const submitTransferForm = useCallback(
+    async (event: FormEvent<HTMLFormElement>) => {
+      event.preventDefault()
+      setSaving(true)
+      setSaveError(null)
+      setSaveNotice(null)
+      try {
+        await submitFormData(new FormData(event.currentTarget))
+        window.location.reload()
+      } catch (error) {
+        setSaveError(error instanceof Error ? error.message : 'Failed to update transfer invitation. Please try again.')
+      } finally {
+        setSaving(false)
+      }
+    },
+    [submitFormData],
+  )
+
   const resetForm = useCallback(() => {
     setFormState(savedFormState)
   }, [savedFormState])
@@ -2230,6 +2248,7 @@ const toggleOrganizationServer = useCallback((serverId: string) => {
         onReassign={handleReassign}
         reassignError={reassignError}
         reassigning={reassigning}
+        onSubmitTransferForm={submitTransferForm}
         onDeleteAgent={confirmDeleteAgent}
         deleteError={deleteError}
       />
@@ -3503,6 +3522,7 @@ type ActionsSectionProps = {
   onReassign: (targetOrgId: string | null) => Promise<void>
   reassignError: string | null
   reassigning: boolean
+  onSubmitTransferForm: (event: FormEvent<HTMLFormElement>) => void
   onDeleteAgent: () => void
   deleteError: string | null
 }
@@ -3519,6 +3539,7 @@ function ActionsSection({
   onReassign,
   reassignError,
   reassigning,
+  onSubmitTransferForm,
   onDeleteAgent,
   deleteError,
 }: ActionsSectionProps) {
@@ -3610,7 +3631,7 @@ function ActionsSection({
                 </p>
                 <p className="text-xs text-indigo-700 mt-1">They'll need to sign in with that email to accept.</p>
               </div>
-              <form method="post" action={urls.detail} className="flex">
+              <form method="post" action={urls.detail} className="flex" onSubmit={onSubmitTransferForm}>
                 <input type="hidden" name="csrfmiddlewaretoken" value={csrfToken} />
                 <input type="hidden" name="action" value="cancel_transfer_invite" />
                 <button type="submit" className={embedded ? 'inline-flex items-center gap-2 rounded-lg border border-slate-200/25 bg-slate-900/35 px-4 py-2 text-sm font-medium text-slate-100 transition-colors hover:border-slate-100/35 hover:bg-slate-900/55' : 'inline-flex items-center gap-2 px-4 py-2 text-sm font-medium text-slate-600 bg-white border border-slate-200 rounded-lg hover:bg-slate-50'}>
@@ -3619,7 +3640,7 @@ function ActionsSection({
               </form>
             </div>
           ) : (
-            <form method="post" action={urls.detail} className="space-y-4">
+            <form method="post" action={urls.detail} className="space-y-4" onSubmit={onSubmitTransferForm}>
               <input type="hidden" name="csrfmiddlewaretoken" value={csrfToken} />
               <input type="hidden" name="action" value="transfer_agent" />
               <div>
