+++ frontend/src/screens/EvalsDetailScreen.tsx
@@ -2,8 +2,7 @@ import { useEffect, useMemo, useState } from 'react'
 import { AlertTriangle, Beaker, Loader2, RefreshCcw, ArrowLeft, Clock, HelpCircle, ChevronDown, ChevronRight, Cpu, BarChart3, Stethoscope } from 'lucide-react'
 
 import { fetchSuiteRunDetail, updateSuiteRunType, fetchRunComparison, fetchSuiteRunComparison, type EvalRun, type EvalSuiteRun, type EvalTask, type LLMRoutingProfileSnapshot, type LLMProfileTokenRange, type LLMProfileTier, type ComparisonResponse, type SuiteComparisonResponse, type ComparisonGroupBy } from '../api/evals'
-import { StatusBadge } from '../components/common/StatusBadge'
-import { RunTypeBadge } from '../components/common/RunTypeBadge'
+import { RunTypeBadge, StatusBadge } from '../components/evals/Badges'
 import { CompareModal, CompareResultsView, type CompareConfig } from '../components/evals'
 
 const formatCurrency = (value?: number | null, digits = 4) => {
