+++ util/integrations.py
@@ -24,9 +24,6 @@ def stripe_status() -> IntegrationStatus:
     return _status("STRIPE_ENABLED", "STRIPE_DISABLED_REASON")
 
 
-def mailgun_status() -> IntegrationStatus:
-    return _status("MAILGUN_ENABLED")
-
 
 def postmark_status() -> IntegrationStatus:
     return _status("POSTMARK_ENABLED")
@@ -55,17 +52,8 @@ def pipedream_status() -> IntegrationStatus:
     )
 
 
-def ensure_stripe_enabled() -> None:
-    status = stripe_status()
-    if not status.enabled:
-        raise IntegrationDisabledError(status.reason or "Stripe integration disabled")
-
 
 def twilio_verify_available() -> bool:
     return bool(getattr(settings, "TWILIO_ENABLED", False) and getattr(settings, "TWILIO_VERIFY_CONFIGURED", False))
 
 
-def postmark_simulation_active() -> bool:
-    if postmark_status().enabled:
-        return False
-    return bool(getattr(settings, "SIMULATE_EMAIL_DELIVERY", False))
